fixed ts errors/ dex/map files
